requirements.txt
django
djangorestframework
django-cors-headers
pillow
psycopg2-binary
djangorestframework-simplejwt
